get_latest_core_version
