+++ run.py
@@ -76,8 +76,8 @@ from voicevox_engine.utility import (
     connect_base64_waves,
     delete_file,
     engine_root,
-    parse_core_version,
 )
+from voicevox_engine.utility.core_version import get_latest_core_version
 
 
 def b64encode_str(s):
@@ -1176,9 +1176,7 @@ if __name__ == "__main__":
         load_all_models=args.load_all_models,
     )
     assert len(synthesis_engines) != 0, "音声合成エンジンがありません。"
-    latest_core_version = str(
-        max([parse_core_version(ver) for ver in synthesis_engines])
-    )
+    latest_core_version = get_latest_core_version(versions=synthesis_engines.keys())
 
     cancellable_engine = None
     if args.enable_cancellable_synthesis:
