+++ voicevox_engine/utility/__init__.py
@@ -1,16 +1,16 @@
+from . import core_version
 from .connect_base64_waves import (
     ConnectBase64WavesException,
     connect_base64_waves,
     decode_base64_waves,
 )
-from .parse_core_version import parse_core_version
 from .path_utility import delete_file, engine_root, get_save_dir
 
 __all__ = [
     "ConnectBase64WavesException",
     "connect_base64_waves",
     "decode_base64_waves",
-    "parse_core_version",
+    "core_version",
     "delete_file",
     "engine_root",
     "get_save_dir",
