--- voicevox_engine/utility/parse_core_version.py
@@ -1,8 +1,0 @@
-from typing import Union
-
-from packaging.version import Version
-from packaging.version import parse as parse_version
-
-
-def parse_core_version(version: str) -> Version:
-    return parse_version(version)
